+++ dev-tools/_exporters/__init__.py
@@ -2,15 +2,19 @@ from ._interface import StructurizrWorkspaceExporter
 from ._interface import ExportedWorkspace
 from ._interface import ExportResult
 from ._interface import ExportFailure
-from ._structurizr_cli import StructurizrCli
-from ._structurizr_lite import StructurizrLite
+from ._structurizr_cli import StructurizrCliForLiteVersion
+from ._structurizr_cli import StructurizrCliForStandaloneVersion
+from ._structurizr_lite import StructurizrLiteForLiteVersion
+from ._structurizr_lite import StructurizrLiteForStandaloneVersion
 
 
 __all__ = [
     "ExportedWorkspace",
     "ExportFailure",
     "ExportResult",
-    "StructurizrCli",
+    "StructurizrCliForLiteVersion",
+    "StructurizrCliForStandaloneVersion",
     "StructurizrWorkspaceExporter",
-    "StructurizrLite",
+    "StructurizrLiteForLiteVersion",
+    "StructurizrLiteForStandaloneVersion",
 ]