+++ dev-tools/tests/test_syntax_plugin.py
@@ -4,7 +4,7 @@ import json
 import logging
 from pathlib import Path
 import tempfile
-from typing import Final, Iterable, Iterator
+from typing import Final, Iterable, Iterator, assert_never
 import pytest
 import _exporter_factory
 import _exporters
@@ -12,9 +12,17 @@ import _logging_tools
 import _exporter_release
 import _cached_downloader
 
+from .helpers import PatternSyntaxPluginDistributive
+
 
 _CUR_DIR_PATH: Final = Path(__file__).parent
 _DOWNLOAD_CACHE_PATH: Final = _CUR_DIR_PATH / ".." / ".cache"
+_JWEAVER_RELEASES: Final = (
+    _exporter_factory.JWeaverRelease(
+        url="https://repo1.maven.org/maven2/org/aspectj/aspectjweaver/1.9.24/aspectjweaver-1.9.24.jar",
+        version="1.9.24",
+    )
+,)
 
 @dataclasses.dataclass(frozen=True, slots=True)
 class SuccessTestResult:
@@ -29,14 +37,31 @@ class FailedTestResult:
 @dataclasses.dataclass(frozen=True, slots=True)
 class TestConfiguration:
     name: str
-    release: _exporter_release.ExporterRelease
+    exporter_config: _exporter_factory.ExporterConfig
     result: SuccessTestResult | FailedTestResult
     workspace_path: Path
 
     @property
     def param_id(self) -> str:
-        return f"release={self.release}, test_case='{self.name}'"
+        config_params = {
+            "release": self.exporter_config.exporter_release.version,
+            "test_case": self.name,
+        }
 
+        match self.exporter_config:
+            case _exporter_factory.LiteVersionExporterConfig():
+                config_params["plugin_version"] = "lite"
+                config_params["jweaver"] = self.exporter_config.jweaver_release.version
+            case _exporter_factory.StandaloneVersionExporterConfig():
+                config_params["plugin_version"] = "standalone"
+
+        match self.exporter_config.exporter_release:
+            case _exporter_release.StructurizrCliRelease():
+                config_params["exporter_type"] = "structurizr_cli"
+            case _exporter_release.StructurizrLiteRelease():
+                config_params["exporter_type"] = "structurizr_lite"
+
+        return " ".join(f"{key}:{value}" for key, value in config_params.items())
 
 @dataclasses.dataclass(frozen=True, slots=True)
 class ReducedTestConfiguration:
@@ -58,16 +83,20 @@ def _create_exporter(exporter_factory: _exporter_factory.ExporterFactory, java_p
         exporter.close()    
 
 
-def _get_test_configs(releases: Iterable[_exporter_release.ExporterRelease], reduced_test_configs: Iterable[ReducedTestConfiguration]) -> list[TestConfiguration]:
+def _get_test_configs(
+    releases: Iterable[_exporter_release.ExporterRelease],
+    reduced_test_configs: Iterable[ReducedTestConfiguration],
+) -> list[TestConfiguration]:
     return [
         TestConfiguration(
             name=reduced_test_config.name,
-            release=release,
+            exporter_config=_exporter_factory.StandaloneVersionExporterConfig(release) if jweaver_release is None else _exporter_factory.LiteVersionExporterConfig(release, jweaver_release),
             result=reduced_test_config.result,
             workspace_path=reduced_test_config.workspace_path,
         )
         for release in releases
         for reduced_test_config in reduced_test_configs
+        for jweaver_release in (*_JWEAVER_RELEASES, None)
     ]
 
 
@@ -173,7 +202,7 @@ def _get_test_configs(releases: Iterable[_exporter_release.ExporterRelease], red
 )
 def test_syntax_plugin(
     test_config: TestConfiguration,
-    syntax_plugin_path: Path,
+    syntax_plugin_dist: PatternSyntaxPluginDistributive,
     java_path: Path,
     samples_dir_path: Path,
     datadir: Path,
@@ -182,8 +211,16 @@ def test_syntax_plugin(
     downloader = _cached_downloader.CachedDownloader(log, _DOWNLOAD_CACHE_PATH)
     workspace_path = samples_dir_path / test_config.workspace_path
 
+    match test_config.exporter_config:
+        case _exporter_factory.LiteVersionExporterConfig():
+            syntax_plugin_path = syntax_plugin_dist.lite_version
+        case _exporter_factory.StandaloneVersionExporterConfig():
+            syntax_plugin_path = syntax_plugin_dist.standalone_version
+        case _:
+            raise assert_never(test_config.exporter_config)
+
     with tempfile.TemporaryDirectory() as temp_dir:
-        exporter_factory = _exporter_factory.get_exporter_factory(downloader, test_config.release, Path(temp_dir), log)
+        exporter_factory = _exporter_factory.get_exporter_factory(downloader, test_config.exporter_config, Path(temp_dir), log)
 
         with _logging_tools.log_action(log, "Run integration test"):
             with _create_exporter(exporter_factory, java_path, syntax_plugin_path) as exporter: