+++ pattern-syntax-plugin/src/main/java/io/github/nifacy/c4patterns/syntax/plugin/aspects/PluginDslContextAspect.java
@@ -1,0 +1,78 @@
+package io.github.nifacy.c4patterns.syntax.plugin.aspects;
+
+import io.github.nifacy.c4patterns.syntax.parser.PatternParser;
+import io.github.nifacy.c4patterns.syntax.parser.PluginCallInfo;
+import io.github.nifacy.c4patterns.syntax.plugin.PatternCallWrapper;
+import org.aspectj.lang.annotation.After;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.Before;
+import org.aspectj.lang.annotation.DeclareParents;
+
+import java.lang.reflect.Field;
+
+@Aspect
+public class PluginDslContextAspect {
+    public static class PatternParserHolderImpl implements PatternParserHolder {
+        private PatternParser patternParser;
+
+        @Override
+        public PatternParser getPatternParser() {
+            return patternParser;
+        }
+
+        @Override
+        public void setPatternParser(PatternParser patternParser) {
+            this.patternParser = patternParser;
+        }
+    }
+
+    @DeclareParents(
+        value = "com.structurizr.dsl.PluginDslContext+", defaultImpl = PatternParserHolderImpl.class
+    )
+    public static PatternParserHolder patternParser;
+
+    @After(
+        """
+                execution(com.structurizr.dsl.PluginDslContext.new(..))
+                && this(ctx)
+                && execution(*.new(..))
+            """
+    )
+    public void afterConstructor(Object ctx) throws Throwable {
+        PatternParser parser = ((PatternParserHolder) ctx).getPatternParser();
+
+        Field fqnField = ctx.getClass().getDeclaredField("fullyQualifiedClassName");
+        fqnField.setAccessible(true);
+        String fullyQualifiedClassName = (String) fqnField.get(ctx);
+
+        if (parser == null && PatternCallWrapper.isWrappedPluginName(fullyQualifiedClassName)) {
+            PatternParser newParser = new PatternParser();
+            newParser.parseHeader(PatternCallWrapper.unwrapFromPluginName(fullyQualifiedClassName));
+            ((PatternParserHolder) ctx).setPatternParser(newParser);
+        }
+    }
+
+    @Before(
+        """
+                execution(* com.structurizr.dsl.PluginDslContext.end(..))
+                && this(ctx)
+            """
+    )
+    public void beforeEnd(Object ctx) throws Throwable {
+        PatternParser parser = ((PatternParserHolder) ctx).getPatternParser();
+
+        if (parser == null) {
+            return;
+        }
+
+        PluginCallInfo info = parser.getPluginCallInfo();
+
+        Field fqnField = ctx.getClass().getDeclaredField("fullyQualifiedClassName");
+        fqnField.setAccessible(true);
+        fqnField.set(ctx, info.getName());
+
+        Field paramsField = ctx.getClass().getDeclaredField("parameters");
+        paramsField.setAccessible(true);
+        paramsField.set(ctx, info.getParameters());
+    }
+}